feat(ocr): extraction quality improvements
Phase 2 of OCR pipeline audit:
- Add extractors for net_amount, vat_amount, sale_date, payment_method, seller_address
- Fix date format ambiguity: context-aware DD.MM.YYYY vs YYYY.MM.DD disambiguation
- Add date range validation (2000-2099) to reject OCR garbage dates
- Replace dangerous "largest amount" fallback with keyword-proximity search
- Add invoice number false-positive filtering (reject NIP/IBAN/account refs)
- Extend IBAN validator to support all EU/international formats (mod-97)
- Change Invoice.amount from float to Decimal for financial precision
- Update all Invoice creation sites to use Decimal


Took 5 minutes

+++ database/models.py
@@ -14,7 +14,7 @@ class Invoice:
     seller_name: str
     invoice_number: str
     invoice_date: date
-    amount: float
+    amount: Decimal
     currency: str = "PLN"
     seller_nip: Optional[str] = None
     bank_account: Optional[str] = None
